Moving SUPPORTED_FIELDS for transformers into serializers
It is only referenced inside of the serializers so we can just
define it there
